Algoritmos de Ordenação

+++ Linguagem_de_Programacao_3/Ex2/Resolução_do_Professor/Calculadora.java
@@ -1,6 +1,6 @@
 public class Calculadora {
-    private double num1;
-    private double  num2;
+    private int num1;
+    private int  num2;
     private char op;
